Added option to give items as an array
