+++ src/components/Dropdown/DropdownComponentClick.vue
@@ -5,10 +5,15 @@
       class="dropbtn"
       @click="toggleVisibility"  
     >
-      Dropdown
+      {{Heading}}
     </button>
     <div ref="dropdown" :class="[isVisible ? isActive : dropdowncontent ]">
-      <slot></slot> 
+      <a v-for="(item, index) in content" v-bind:key="index" 
+        :href="item.href"
+        @click="click"
+      >
+        {{item.option}} 
+      </a>
     </div>
   </div>
 </template>
@@ -17,7 +22,13 @@
 import { ref } from 'vue';
 
 export default {
-    setup() {
+      props: {
+        Heading: String,
+    content: Array,
+    href: String,
+  }, 
+  setup(props) {
+    const url = ref(props.href) 
         const isVisible = ref(false);
       const isActive = ref("isActive"); 
       const dropdowncontent = ref("dropdown-content");
@@ -27,12 +38,18 @@ export default {
       console.log(isVisible.value);
     }  
 
+    function click () {
+      window.open(this.$refs.dropdown.children[this.focusedIndex].href,"_self") 
+    }
+
     return {
-        isVisible,
-        isActive,
+      isVisible,
+      url,
+      isActive,
         dropdowncontent,
         toggleVisibility, 
-    }
+      click
+    } 
     } 
 };
 </script>
@@ -69,5 +86,18 @@ export default {
 } 
 .dropdown:hover .dropbtn {
   background-color: #2f382a;
+}
+a {
+  color: 	white;
+  padding: 12px 16px;
+  font-family: sans-serif;
+  text-decoration: none;
+  display: block;
+  background-color: #475447;
+} 
+
+a:hover {
+  background-color: white;
+  color: #2f382a;
 }
 </style>
